+++ NEWS.md
@@ -1,6 +1,15 @@
 Version 0.9
 ---------------------------------------------------------------------
 
+- Update all references to point to DeltaRho (0.9.7)
+- vdbConvert method to convert vdb to be compatible with TrelliscopeJS (0.9.7)
+- Add thumbIndex to makeDisplay (0.9.7)
+- Fix png and jpeg imports (0.9.7)
+- reveal long cog values in tabel and label views and put max width on table cog labels (0.9.6)
+- preserve nrow and ncol in state layout hash (0.9.6)
+- add option to view() to not copy viewer files (0.9.6)
+- fix bug with htmlwidget related displays not showing (0.9.6)
+- add 'base64png' panel function type (0.9.6)
 - fix state spec in `qtrellis()` (0.9.5)
 - don't run scagnostics example since it uses a suggested package (0.9.5)
 - add favicon.ico (0.9.4)
@@ -204,5 +213,3 @@ BUG FIXES
 - fix arrow key shortcuts to only operate when appropriate
 - fix thumbnail generation to yield small files to speed up viewer startup
 - fix robustness issues with page input in viewer
-
-